refactor(ai-tracing): consolidate span operations in ModelSpanTracker (#9151)
## Description

Centralized all LLM generation span operations within ModelSpanTracker
to eliminate redundant llmAISpan parameter passing.

**Changes:**

- **ModelSpanTracker API**: Added public methods for span operations
  - `getTracingContext()`: Returns current step span or model span
  - `reportGenerationError()`: Report errors on generation span
  - `endGeneration()`: End generation span with output/attributes
  - `updateGeneration()`: Update generation span
  - Made internal methods private (startStepSpan, endStepSpan, etc.)

- **Removed llmAISpan parameter**: No longer passed through call chain
  - Removed from LoopOptions, LoopRun, OuterLLMRun types
  - Removed from loop(), workflowLoopStream() signatures
  - llmAISpan now only used locally in model.loop.ts

- **Updated call sites**: Use ModelSpanTracker methods instead
  - model.loop.ts: Use reportGenerationError() and endGeneration()
  - loop.ts, stream.ts: Use getTracingContext() for tracing context
  - tool-call-step.ts: Use getTracingContext() for tool spans

**Benefits:**
- Single source of truth for generation span operations
- Cleaner API with fewer parameters
- Better encapsulation of span lifecycle
- Consistent pattern for accessing tracing context


## Related Issue(s)

#6773

## Type of Change

- [ ] Bug fix (non-breaking change that fixes an issue)
- [ ] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)
- [ ] Documentation update
- [X] Code refactoring
- [ ] Performance improvement
- [ ] Test update

## Checklist

- [ ] I have made corresponding changes to the documentation (if
applicable)
- [ ] I have added tests that prove my fix is effective or that my
feature works

---------
