+++ packages/core/src/ai-tracing/index.ts
@@ -11,4 +11,4 @@ export * from './utils';
 export * from './context';
 export * from './exporters';
 export * from './span_processors';
-export * from './chunk-tracing';
+export * from './model-tracing';
