+++ packages/core/src/llm/model/model.loop.ts
@@ -228,7 +228,6 @@ export class MastraLLMVNext extends MastraBase {
         structuredOutput,
         outputProcessors,
         returnScorerData,
-        llmAISpan,
         modelSpanTracker,
         requireToolApproval,
         agentId,
@@ -257,7 +256,7 @@ export class MastraLLMVNext extends MastraBase {
                 },
                 e,
               );
-              llmAISpan?.error({ error: mastraError });
+              modelSpanTracker?.reportGenerationError({ error: mastraError });
               this.logger.trackException(mastraError);
               throw mastraError;
             }
@@ -303,12 +302,12 @@ export class MastraLLMVNext extends MastraBase {
                 },
                 e,
               );
-              llmAISpan?.error({ error: mastraError });
+              modelSpanTracker?.reportGenerationError({ error: mastraError });
               this.logger.trackException(mastraError);
               throw mastraError;
             }
 
-            llmAISpan?.end({
+            modelSpanTracker?.endGeneration({
               output: {
                 files: props?.files,
                 object: props?.object,
@@ -361,7 +360,7 @@ export class MastraLLMVNext extends MastraBase {
         },
         e,
       );
-      llmAISpan?.error({ error: mastraError });
+      modelSpanTracker?.reportGenerationError({ error: mastraError });
       throw mastraError;
     }
   }
