+++ packages/core/src/loop/test-utils/options.ts
@@ -27,6 +27,7 @@ import {
   testUsage2,
   createMessageListWithUserMessage,
 } from './utils';
+import { ModelSpanTracker } from '../../ai-tracing';
 
 export function optionsTests({ loopFn, runId }: { loopFn: typeof loop; runId: string }) {
   describe('options.abortSignal', () => {
@@ -35,6 +36,7 @@ export function optionsTests({ loopFn, runId }: { loopFn: typeof loop; runId: st
 
       const abortController = new AbortController();
       const toolExecuteMock = vi.fn().mockResolvedValue('tool result');
+      const modelSpanTracker = new ModelSpanTracker();
 
       const result = loopFn({
         runId,
@@ -64,6 +66,7 @@ export function optionsTests({ loopFn, runId }: { loopFn: typeof loop; runId: st
           abortSignal: abortController.signal,
         },
         agentId: 'agent-id',
+        modelSpanTracker,
       });
 
       await convertAsyncIterableToArray(result.aisdk.v5.fullStream as any);
